Fix rotary embedding

+++ ochat/models/unpadded_gemma.py
@@ -97,7 +97,7 @@ class UnpaddedGemmaRotaryEmbedding(torch.nn.Module):
 
         # RoPE
         inv_freq = 1.0 / (base ** (torch.arange(0, dim, 2, dtype=torch.int64, device=device).float() / dim))
-        t = torch.arange(max_position_embeddings, dtype=torch.float32, device=device)
+        t = torch.arange(max_position_embeddings, dtype=torch.int64, device=device).type_as(inv_freq)
         freqs = torch.outer(t, inv_freq)
 
         # Different from paper, but it uses a different permutation in order to obtain the same calculation
